fixing filters

+++ src/controller/static_mates.controller.ts
@@ -1,3 +1,5 @@
+import pkg from "lodash";
+const { difference } = pkg;
 import { StaticMate } from "../models/static_mates.model.js";
 import {
   createStaticMatePayload,
@@ -8,19 +10,31 @@ import {
 } from "./types/static_mate_types.js";
 import { Request, Response } from "express"; // Make sure you are importing from express
 import { errorMonitor } from "stream";
+import { deleteFloorPayload } from "./types/raid_floor_types.js";
 
 ///////Read or Pull all
 export const findAllStaticMates = async (
   req: Request,
   res: Response
 ): Promise<StaticMateServiceReturn> => {
-
   try {
     //finaAllStaticMates = static_name, raid_member_role
     const getAllStaticMates = req.query;
+    const default_fields: string[] = Object.keys(findAllStaticMates);
+    const REQUIRED_STATIC_MATE_FIELDS: string[] = ["name", "role"];
+    const missingFields: string[] = difference(
+      REQUIRED_STATIC_MATE_FIELDS,
+      default_fields
+    );
+    if (missingFields.length! == 0) {
+      return {
+        status: "Error",
+        data: "Missing required fields:",
+      };
+    }
     console.log("Calling find all for Static Mates");
     const getStaticMates: StaticMate[] = await StaticMate.findAll({
-      where: getAllStaticMates,  //static_name, raid_member_role
+      where: getAllStaticMates, //static_name, raid_member_role
       order: [["id", "asc"]],
     });
     res.status(200).json({
@@ -50,29 +64,27 @@ export const findAllStaticMates = async (
 
 //Pull a SINGLE PERSON NEEDS TO BE FIXED
 export const findSingleStaticMate = async (
-  req: Request<{ id: number }>,
+  req: Request<{ id: number }, {}, {}, createStaticMatePayload>,
   res: Response
 ): Promise<StaticMateServiceReturn> => {
   try {
-    const getSingleStaticMember = req.query;
+
+
     console.log("Calling find one Static Mates");
     const { id }: { id: number } = req.params;
-    const find_single_static_mate: StaticMate = await StaticMate.findOne(
-      {
-        where: {
-          id: id,
-        },
-      }
-    );
+    const find_single_static_mate: StaticMate = await StaticMate.findOne({
+      where: {
+        id: id,
+      },
+    });
     res.status(200).json({
-        status: "Sucess",
-        data: find_single_static_mate,
+      status: "Sucess",
+      data: find_single_static_mate,
     });
     return {
       status: "Sucess",
       data: find_single_static_mate,
     };
-
   } catch (error) {
     console.error(error);
     res.status(500).json({
@@ -93,25 +105,26 @@ export const findSingleStaticMate = async (
 
 //adding object to createSingleStaticMate is acting weird
 export const createSingleStaticMate = async (
-  req: Request<{ createStaticMatePayload }>,
+  req: Request<{}, createStaticMatePayload>,
   res: Response
 ): Promise<StaticMateServiceReturn> => {
   try {
+    //finaAllStaticMates = static_name, raid_member_role
+    console.log("calling createSingleStaticMate")
+    const default_fields: string[] = Object.keys(req.body);
+    console.log (default_fields);
+    const REQUIRED_STATIC_MATE_FIELDS: string[] = ["name", "role"];
+    const missingFields: string[] = difference(
+      REQUIRED_STATIC_MATE_FIELDS,
+      default_fields
+    );
+    if (missingFields.length! == 0) {
+      return {
+        status: "Error",
+        data: "Missing required fields:",
+      };
+    }
     //properties on the query object and the query object is property on the request object
-    const createSingleStaticMate = req.query;
-    const default_fields: string[] = Object.keys(createSingleStaticMate);
-    const REQUIRED_STATIC_MATE_FIELDS: string[] =[
-      'name',
-      'role'
-    ];
-    const missingFields: string [] = difference(REQUIRED_STATIC_MATE_FIELDS, default_fields);
-    if (missingFields.length ! == 0){
-      return{
-        status: 'Error',
-        data: 'Missing required fields:' 
-      }
-    };
-
     const createStaticMatePayload: createStaticMatePayload = req.body;
     console.log("req.body", req.body);
 
@@ -130,7 +143,6 @@ export const createSingleStaticMate = async (
     return {
       status: "Sucess",
       data: single_static_mate,
-
     };
   } catch (error) {
     res.status(500).json({
@@ -148,25 +160,35 @@ export const createSingleStaticMate = async (
 
 /////Update
 export const updateStaticMate = async (
-  req: Request<{ id: number }, any, updateStaticMatePayload>,
+  req: Request<{ id: number }, updateStaticMatePayload>,
   res: Response
 ): Promise<StaticMateServiceReturn> => {
   try {
+    //finaAllStaticMates = static_name, raid_member_role
+    const default_fields: string[] = Object.keys(updateStaticMate);
+    const REQUIRED_STATIC_MATE_FIELDS: string[] = ["name", "role"];
+    const missingFields: string[] = difference(
+      REQUIRED_STATIC_MATE_FIELDS,
+      default_fields
+    );
+    if (missingFields.length! == 0) {
+      return {
+        status: "Error",
+        data: "Missing required fields:",
+      };
+    }
     //do a find first to see if that thing exist
     //also ends things early
-    const updateStaticMate = req.query;
     const { id }: { id: number } = req.params;
     const updateStaticMateInfo: updateStaticMatePayload = req.body;
     console.log("req.body is reading:", updateStaticMateInfo);
-    const [affectedCount, affectedRows]: [number, StaticMate[]] = await StaticMate.update(
-      updateStaticMateInfo,
-      {
+    const [affectedCount, affectedRows]: [number, StaticMate[]] =
+      await StaticMate.update(updateStaticMateInfo, {
         where: {
           id: id,
         },
         returning: true,
-      }
-    );
+      });
 
     res.status(200).json({
       message: {
@@ -178,7 +200,6 @@ export const updateStaticMate = async (
       status: "Sucess",
       data: affectedRows,
     };
-
   } catch (error) {
     console.error(error);
     res.status(500).json({
@@ -198,11 +219,23 @@ export const updateStaticMate = async (
 ///////Delete
 //export const deleteStaticMate = async (req: Request<(id: string), any, deleteStaticMatePayload>, res:Response): Promise<StaticMateServiceReturn> => {
 export const deleteStaticMate = async (
-  req: Request<{ id: string }>,
+  req: Request<{ id: string }, {}, {}, deleteStaticMatePayload>,
   res: Response
 ): Promise<StaticMateServiceReturn> => {
   try {
-    const deleteStaticMate = req.query;
+    //finaAllStaticMates = static_name, raid_member_role
+    const default_fields: string[] = Object.keys(findSingleStaticMate);
+    const REQUIRED_STATIC_MATE_FIELDS: string[] = ["name", "role"];
+    const missingFields: string[] = difference(
+      REQUIRED_STATIC_MATE_FIELDS,
+      default_fields
+    );
+    if (missingFields.length! == 0) {
+      return {
+        status: "Error",
+        data: "Missing required fields:",
+      };
+    }
     const { id }: { id: string } = req.params;
     const delete_static_mate: number = await StaticMate.destroy({
       where: {
@@ -238,14 +271,13 @@ export const deleteStaticMate = async (
 //updateStaticMate({query: {id: 4 ,static_name: "Higgs", raid_member_role: "Ranged DPS"}}, {})
 //the second half is the response so you can chain controllers to each other the second object is also a res
 
-
 // for(const key in getAllStaticMates) {
 //   if(!unit_keys.includes(key) {
 //   return {
 //   status: 'Error',
 //   data:  Invalid Field ${keys}
-  
+
 // }
 //   }
-  
+
 //   }
